test(#4196): split over-threshold corruption-corpus test file — new issue_4196_salvage_oom_bounds.rs target + shared support/salvage_corpus.rs
Campsite rule (epic #1135): issue_4196_salvage_corruption_corpus.rs had
crossed the ~1500-line test-file threshold (1542 lines, flagged as a
round-11 process note). Lead-mandated split before the round-12 review, a
pure move -- no behavior changed:

- issue_4196_salvage_corruption_corpus.rs (686 lines): the ORIGINAL
  refusal/classification cases -- chunk-CRC flip, damaged boundary source,
  the three ComponentUnreadable refusal tests, and
  swapped_index_entry_keys_classify_key_mismatch.
- issue_4196_salvage_oom_bounds.rs (771 lines, new target): the 4
  OOM/allocation-bound regression tests spanning rounds 9-11, grouped
  together since they exercise the same chunk-range/pre-flight code area
  and share the round-11 audit's evidence trail.
- tests/support/salvage_corpus.rs (187 lines, new): the shared
  fixture-resolution + Index.db-structural helpers both targets depend on,
  cross-referenced via the same super::datasets_root pattern
  support/header_refusal.rs already establishes.

New target named in the gate's write-tests list next to its sibling
issue_4196_salvage_corruption_corpus (#3522). Confirmed the split preserves
the test count exactly: 10 tests in one file before -> 6 + 4 after, none
lost or duplicated.

Re-verified: cargo fmt --check clean; clippy clean (lib + all four salvage
test targets); cargo test -p cqlite-core --lib --features write-support
4060 passed (unchanged); cargo test -p cqlite-cli --lib --features
write-support 233 passed (unchanged); all 15 salvage cqlite-core tests
(6+4+4+1, same total as the 10+4+1 pre-split) and CLI tests (8) pass
against the real corpus; test_salvage_no_resync_scan.sh passes;
features-load-bearing 61/61 (1712 source files scanned, was 1710).
